Fixed demo bug
Was an issue with the code i was using to get the ids, the queries to do so included spaces and so wouldnt work in a url. To be honest i dont know how it was ever working in the first place. Changed it to a simpler way by getting the name and id at once and returning in a recipe list instead of a string list made up of just the names which I used to get the ids in another function